+++ CHANGES.rst
@@ -1,6 +1,7 @@
 1.1.2 - Unreleased
 ==================
 
+- provide default and by-convention assignment of fabfiles
 - fix string escapes for geli setup in rc.conf
 - make sshd listen address configurable
 - fix permission of periodic scripts in zfs_auto_snapshot role